requirements.txt
yt-dlp
openai-whisper
transformers
accelerate
sentencepiece